+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 - Added `EntityReference`:
   An entity field wrapper that allows lazy load of sub-entities.
 - Added `EntityResolutionRules` to allow lazy or eager selects.
-- reflection_factory: ^1.2.8
+- `EntityProvider`:
+  - `getEntityByID`: Added parameter `sync`.
+- reflection_factory: ^1.2.9
 
 ## 1.3.6